Fixing Error on Thermostat Update when thermostat has no fan
If the current thermostat has no fan, we need to check if thermostat
has a fan before setting the fanTimerActive value.  If this value is
posted to the api and the thermostat does not have a fan, then the
error below is received.

https://developers.nest.com/documentation/cloud/error-messages#no-hvac-f
an

Must have been coded by someone in a hotel climate than I :)